move all functions related to presenting data excluding console output to csbench_visualize.c
